move artifact type to items
