+++ game/magic/artifact/item.go
@@ -3,23 +3,388 @@ package artifact
 import (
 	"bytes"
 	"fmt"
+    "slices"
 	_ "log"
 
-	"github.com/kazzmir/master-of-magic/game/magic/data"
 	"github.com/kazzmir/master-of-magic/lib/lbx"
+	"github.com/kazzmir/master-of-magic/game/magic/data"
+    "github.com/kazzmir/master-of-magic/game/magic/spellbook"
 )
 
-// TODO: use Artifact
-type Item struct {
-	Name string
-	Image int
-	Slot ArtifactSlot
-	Type ArtifactType
-	Powers []Power
+type ArtifactType int
+const (
+    ArtifactTypeNone ArtifactType = iota
+    ArtifactTypeSword
+    ArtifactTypeMace
+    ArtifactTypeAxe
+    ArtifactTypeBow
+    ArtifactTypeStaff
+    ArtifactTypeWand
+    ArtifactTypeMisc
+    ArtifactTypeShield
+    ArtifactTypeChain
+    ArtifactTypePlate
+)
+
+func (a ArtifactType) Name() string {
+    switch a {
+        case ArtifactTypeSword: return "Sword"
+        case ArtifactTypeMace: return "Mace"
+        case ArtifactTypeAxe: return "Axe"
+        case ArtifactTypeBow: return "Bow"
+        case ArtifactTypeStaff: return "Staff"
+        case ArtifactTypeWand: return "Wand"
+        case ArtifactTypeMisc: return "Misc"
+        case ArtifactTypeShield: return "Shield"
+        case ArtifactTypeChain: return "Chain"
+        case ArtifactTypePlate: return "Plate"
+    }
+
+    return ""
+}
+
+type ArtifactSlot int
+const (
+    ArtifactSlotMeleeWeapon ArtifactSlot = iota
+    ArtifactSlotRangedWeapon
+    ArtifactSlotMagicWeapon
+    ArtifactSlotAnyWeapon
+    ArtifactSlotArmor
+    ArtifactSlotJewelry
+)
+
+// the index in itemisc.lbx for this slot
+func (slot ArtifactSlot) ImageIndex() int {
+    switch slot {
+        case ArtifactSlotMeleeWeapon: return 19
+        case ArtifactSlotRangedWeapon: return 20
+        case ArtifactSlotMagicWeapon: return 22
+        case ArtifactSlotAnyWeapon: return 21
+        case ArtifactSlotArmor: return 24
+        case ArtifactSlotJewelry: return 23
+    }
+
+    return -1
+}
+
+func (slot ArtifactSlot) CompatibleWith(kind ArtifactType) bool {
+    switch slot {
+        case ArtifactSlotMeleeWeapon:
+            return kind == ArtifactTypeSword || kind == ArtifactTypeMace || kind == ArtifactTypeAxe
+        case ArtifactSlotRangedWeapon:
+            return kind == ArtifactTypeSword || kind == ArtifactTypeMace || kind == ArtifactTypeAxe || kind == ArtifactTypeBow
+        case ArtifactSlotMagicWeapon:
+            return kind == ArtifactTypeStaff || kind == ArtifactTypeWand
+        case ArtifactSlotAnyWeapon:
+            return kind == ArtifactTypeStaff || kind == ArtifactTypeWand || kind == ArtifactTypeSword || kind == ArtifactTypeMace || kind == ArtifactTypeAxe
+        case ArtifactSlotArmor:
+            return kind == ArtifactTypeShield || kind == ArtifactTypeChain || kind == ArtifactTypePlate
+        case ArtifactSlotJewelry:
+            return kind == ArtifactTypeMisc
+    }
+
+    return false
+}
+
+type Power interface {
+    String() string
+    Cost() int
+    GetAmount() int
+}
+
+type PowerAttack struct {
+    Amount int
+}
+
+func (p *PowerAttack) GetAmount() int {
+    return p.Amount
+}
+
+func (p *PowerAttack) Cost() int {
+    switch p.Amount {
+        case 1: return 50
+        case 2: return 100
+        case 3: return 200
+        case 4: return 350
+        case 5: return 550
+        case 6: return 800
+    }
+
+    return 800
+}
+
+func (p *PowerAttack) String() string {
+    return fmt.Sprintf("+%v Attack", p.Amount)
+}
+
+type PowerDefense struct {
+    Amount int
+}
+
+func (p *PowerDefense) GetAmount() int {
+    return p.Amount
+}
+
+func (p *PowerDefense) Cost() int {
+    switch p.Amount {
+        case 1: return 50
+        case 2: return 100
+        case 3: return 200
+        case 4: return 350
+        case 5: return 550
+        case 6: return 800
+    }
+
+    return 800
+}
+
+func (p *PowerDefense) String() string {
+    return fmt.Sprintf("+%v Defense", p.Amount)
+}
+
+type PowerToHit struct {
+    Amount int
+}
+
+func (p *PowerToHit) GetAmount() int {
+    return p.Amount
+}
+
+func (p *PowerToHit) Cost() int {
+    switch p.Amount {
+        case 1: return 400
+        case 2: return 800
+        case 3: return 1200
+    }
+
+    return 1200
+}
+
+func (p *PowerToHit) String() string {
+    return fmt.Sprintf("+%v To Hit", p.Amount)
+}
+
+type PowerSpellSkill struct {
+    Amount int
+}
+
+func (p *PowerSpellSkill) GetAmount() int {
+    return p.Amount
+}
+
+func (p *PowerSpellSkill) Cost() int {
+    switch p.Amount {
+        case 5: return 200
+        case 10: return 400
+        case 15: return 800
+        case 20: return 1600
+    }
+
+    return 1600
+}
+
+func (p *PowerSpellSkill) String() string {
+    return fmt.Sprintf("+%v Spell Skill", p.Amount)
+}
+
+type PowerSpellSave struct {
+    Amount int
+}
+
+func (p *PowerSpellSave) GetAmount() int {
+    return p.Amount
+}
+
+func (p *PowerSpellSave) Cost() int {
+    switch p.Amount {
+        case -1: return 100
+        case -2: return 200
+        case -3: return 400
+        case -4: return 800
+    }
+
+    return 800
+}
+
+func (p *PowerSpellSave) String() string {
+    return fmt.Sprintf("%v Spell Save", p.Amount)
+}
+
+type PowerResistance struct {
+    Amount int
+}
+
+func (p *PowerResistance) GetAmount() int {
+    return p.Amount
+}
+
+func (p *PowerResistance) Cost() int {
+    switch p.Amount {
+        case 1: return 50
+        case 2: return 100
+        case 3: return 200
+        case 4: return 350
+        case 5: return 550
+        case 6: return 800
+    }
+
+    return 800
+}
+
+func (p *PowerResistance) String() string {
+    return fmt.Sprintf("+%v Resistance", p.Amount)
+}
+
+func (p *PowerMovement) Cost() int {
+    switch p.Amount {
+        case 1: return 100
+        case 2: return 200
+        case 3: return 400
+        case 4: return 800
+    }
+
+    return 800
+}
+
+func (p *PowerMovement) GetAmount() int {
+    return p.Amount
+}
+
+type PowerMovement struct {
+    Amount int
+}
+
+func (p *PowerMovement) String() string {
+    return fmt.Sprintf("+%v Movement", p.Amount)
+}
+
+type PowerSpellCharges struct {
+    Spell spellbook.Spell
+    Charges int
+}
+
+func (p *PowerSpellCharges) GetAmount() int {
+    return 0
+}
+
+func (p *PowerSpellCharges) Cost() int {
+    // FIXME: depends on the spell and the charges
+    return 0
+}
+
+func (p *PowerSpellCharges) String() string {
+    return "Spell Charges"
+}
+
+type Artifact struct {
+    Type ArtifactType
+    Image int
+    Name string
+    Powers []Power
 	Abilities []data.Ability
 }
 
-func ReadItems(cache *lbx.LbxCache) ([]Item, error) {
+func (artifact *Artifact) HasAbility(ability data.AbilityType) bool {
+    switch ability {
+        case data.AbilityLargeShield: return artifact.Type == ArtifactTypeShield
+        // TODO: handle more enchantments on the artifact, such as vampiric
+    }
+
+    return false
+}
+
+func (artifact *Artifact) AddPower(power Power) {
+    artifact.Powers = append(artifact.Powers, power)
+}
+
+func (artifact *Artifact) RemovePower(remove Power) {
+    artifact.Powers = slices.DeleteFunc(artifact.Powers, func (power Power) bool {
+        return remove == power
+    })
+}
+
+func addPowers[T Power](powers []Power) int {
+    amount := 0
+    for _, power := range powers {
+        convert, ok := power.(T)
+        if ok {
+            amount += convert.GetAmount()
+        }
+    }
+
+    return amount
+}
+
+func (artifact *Artifact) MeleeBonus() int {
+    switch artifact.Type {
+        case ArtifactTypeSword, ArtifactTypeMace, ArtifactTypeAxe, ArtifactTypeMisc:
+            return addPowers[*PowerAttack](artifact.Powers)
+        default:
+            return 0
+    }
+}
+
+func (artifact *Artifact) RangedAttackBonus() int {
+    switch artifact.Type {
+        case ArtifactTypeBow, ArtifactTypeMisc:
+            return addPowers[*PowerAttack](artifact.Powers)
+        default:
+            return 0
+    }
+}
+
+func (artifact *Artifact) DefenseBonus() int {
+    base := addPowers[*PowerDefense](artifact.Powers)
+    switch artifact.Type {
+        case ArtifactTypeChain:
+            base += 1
+        case ArtifactTypePlate:
+            base += 2
+    }
+
+    return base
+}
+
+func (artifact *Artifact) ResistanceBonus() int {
+    return addPowers[*PowerResistance](artifact.Powers)
+}
+
+func (artifact *Artifact) Cost() int {
+    base := 0
+    switch artifact.Type {
+        case ArtifactTypeSword: base = 100
+        case ArtifactTypeMace: base = 100
+        case ArtifactTypeAxe: base = 100
+        case ArtifactTypeBow: base = 100
+        case ArtifactTypeStaff: base = 300
+        case ArtifactTypeWand: base = 200
+        case ArtifactTypeMisc: base = 50
+        case ArtifactTypeShield: base = 100
+        case ArtifactTypeChain: base = 100
+        case ArtifactTypePlate: base = 300
+    }
+
+    powerCost := 0
+    spellCost := 0
+    for _, power := range artifact.Powers {
+        spell, isSpell := power.(*PowerSpellCharges)
+        if isSpell {
+            spellCost += spell.Cost()
+        } else {
+            powerCost += power.Cost()
+        }
+    }
+
+    // jewelry costs are 2x
+    if artifact.Type == ArtifactTypeMisc {
+        powerCost *= 2
+    }
+
+    return base + powerCost + spellCost
+}
+
+
+func ReadItems(cache *lbx.LbxCache) ([]Artifact, error) {
 	itemData, err := cache.GetLbxFile("itemdata.lbx")
 	if err != nil {
 		return nil, fmt.Errorf("unable to read itemdata.lbx: %v", err)
@@ -43,7 +408,7 @@ func ReadItems(cache *lbx.LbxCache) ([]Item, error) {
 		return nil, fmt.Errorf("unsupported itemdata.lbx")
 	}
 
-	var out []Item
+	var out []Artifact
 
 	slotMap := map[byte]ArtifactSlot{
 		1: ArtifactSlotMeleeWeapon,
@@ -126,6 +491,9 @@ func ReadItems(cache *lbx.LbxCache) ([]Item, error) {
 		if !exists {
 			return nil, fmt.Errorf("Invalid slot type %v", slotValue)
 		}
+
+        // TODO: use this somehow? seems unnecessary
+        _ = slot
 
 		// Type
 		typeValue, err := lbx.ReadByte(reader)
@@ -271,15 +639,15 @@ func ReadItems(cache *lbx.LbxCache) ([]Item, error) {
 			return nil, fmt.Errorf("read error: %v", err)
 		}
 
-		item := Item{
+		artifact := Artifact{
 			Name: string(name),
 			Image: int(image),
-			Slot: slot,
+			// Slot: slot,
 			Type: artifactType,
 			Powers: powers,
 			Abilities: abilities,
 		}
-		out = append(out, item)
+		out = append(out, artifact)
 	}
 
 	return out, nil
